backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
supabase-py>=0.0.3
psycopg2-binary>=2.9.6
python-decouple>=3.8
